seperate deployable class
